Finalise the controller template
Make it as easy as possible, the code is doing essentially a call to the Server using the generated SDK, leave a room for flexiblity using the DTO function to convert Rust generate Struct to a OpenAPI model.

Have to think about this approach because it might be redundant, but at the same time make it more flexible because the template which generates the client are often no so easy to customise, snake_case to camelCase etc. For now I'll keep it redundant.