Massive refactoring
Implement x-kubernetes-operator-include-tags to decide which controllers to create.

Simplify the code generation process, use separate binaries for the code generation, and store the code and crd generation in a separate directories. Remove build.rs, it's hard to debug, when I build the binary with symbols I can go through the code easier.

Improve examples generation, if there is an array of examples there will be a manifest generated with multiple manifests, all based on what's declared in OAS.

Decouple controller generation from types generations, controllers will be generated based on the paths, this make it more flexible for different use-cases.

Types are just types and will be used in the controllers. they are not excluded from generation by x-kubernetes-include-tags because they don't have a tag option afaik.