Add progress
Restructure the project, now every domain logic is a part of its own package, started using workspaces.

Added E2E testing setup, where I'm spinning a local kubernetes cluster and deploying the resource and check what's the modifications to the resource are after callinga specific function, i.e add_finalizer.

Fixed an issue with generated unsorted data, where every time running task generate the order was not guaranteed, now it is.

Add package task to allow to package the operator and ship it to a kubernetes cluster.

Fixed RBAC cluster role, allow it to deploy CRDs into the cluster to simplify the process.

Added a way to pass ACCESS_TOKEN to the operator from a kubernetes secret to the environment so the operator can communicate with protected API's.