Keep directory structure simple
Forget about rust convention if it's a cli, then it makes sense to put it in a directory called cli and call this crate cli - easier to understand.

Later I will be publishing this one as a binary that the consumers can download and run on a fresh new project.