feat: Add test-cli-output task to Taskfile.yaml
Add a new task called test-cli-output to Taskfile.yaml. This task is used to test the CLI output. It runs the command 'cargo run --package kopgen init temp'. This task is useful for verifying the correctness of the CLI init output.

Also, remove the temp directory in the clean task in Taskfile.yaml.