chore: Update cross-compilation settings for arm64 linux target in workflow
Unfortunatly, github still don't support managed runners for arm64 :(

It's a sad fact that we're almost in the year 2025 and they only offer it for enterprise customers. :dinosaur

I can run it on my raspberry pi, which is arm64, but it's not ideal to have it self-hosted.

Instead will cross-compile this part for now with musl libc.

Signed-off-by: Eden Reich <eden.reich@gmail.com>